requirements.txt
flet
Pillow=10.3.0
pyinstaller